mmore inputs in parameters in paraminputs component
